Comificaciones de Cuenta
Las cuentas corrientes ya no pueden solicitar prestamos´, los métodos aumentar y disminuir saldo se heredan de la clase Cuenta ya que en CuentaCorriente y CuentaAhorro se implementan de la misma manera. Se corrigieron errores generados en Prestamos, Multa, Pago y Cliente
