+++ src/gestorAplicacion/usuario/Cuenta.java
@@ -12,12 +12,9 @@ public abstract class Cuenta {
     protected int saldoTotal;
     protected int SaldoDisponible;
     protected Cliente titular;
-    protected Prestamo prestamo;
     public ArrayList<Bolsillo> misBolsillos = new ArrayList<>();
-    protected int deuda;
-    protected Multa multa;
     protected boolean estado;
-    
+
     public Cuenta(Cliente titular) {
         this.titular = titular;
         this.estado = true;
@@ -25,21 +22,25 @@ public abstract class Cuenta {
 
     }
 
-    public abstract void aumentarSaldo(int cantidad);
+    public abstract String toString();
 
-    public abstract void disminuirSaldo(int cantidad);
+    public void aumentarSaldo(int cantidad) {
+        if (isEstado()){
+            setSaldoTotal(getSaldoTotal() + cantidad);
+        }
+    }
 
-    public abstract void  solicitarPrestamo(int valorPrestamo, String tipoPrestamo);
+    public void disminuirSaldo(int cantidad) {
+        if (isEstado() && (getSaldoDisponible() > cantidad)){
+            setSaldoDisponible(getSaldoDisponible() - cantidad);
+        }
+    }
 
-    public abstract boolean tieneMultta();
-
-    public abstract boolean tienePrestamo();
-
-    public int getNumero() {
+    public static int getNumero() {
         return numero;
     }
 
-    public void setNumero(int numero) {
+    public static void setNumero(int numero) {
         Cuenta.numero = numero;
     }
 
@@ -67,22 +68,6 @@ public abstract class Cuenta {
         this.titular = titular;
     }
 
-    public Prestamo getPrestamo() {
-        return prestamo;
-    }
-
-    public void setPrestamo(Prestamo prestamo) {
-        this.prestamo = prestamo;
-    }
-
-    public int getDeuda() {
-        return deuda;
-    }
-
-    public void setDeuda(int deuda) {
-        this.deuda = deuda;
-    }
-
     public ArrayList<Bolsillo> getMisBolsillos() {
         return misBolsillos;
     }
@@ -91,39 +76,11 @@ public abstract class Cuenta {
         this.misBolsillos = misBolsillos;
     }
 
-    public Multa getMulta() {
-        return multa;
-    }
-
-    public void setMulta(Multa multa) {
-        this.multa = multa;
-    }
-
     public boolean isEstado() {
         return estado;
     }
 
     public void setEstado(boolean estado) {
         this.estado = estado;
-    }
-
-
-
-    public void setAhorro(ArrayList<Bolsillo> ahorro) {
-        this.misBolsillos = ahorro;
-    }
-
-
-    @Override
-    public String toString() {
-        return "Cuenta{" +
-                "saldoTotal=" + saldoTotal +
-                ", SaldoDisponible=" + SaldoDisponible +
-                ", titular=" + titular +
-                ", prestamo=" + prestamo +
-                ", deuda=" + deuda +
-                ", multa=" + multa +
-                ", estado=" + estado +
-                '}';
     }
 }
